requirements.txt
flask==2.3.2
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.6.1
langchain==0.2.0
langchain-community==0.2.0
langchain-huggingface==0.0.3
langchain-chroma==0.2.0
pymupdf==1.23.0
pytesseract==0.3.10
pdf2image==1.16.3
torch==2.0.1
cachetools==5.3.1